+++ src/libromdata/Media/Wim.cpp
@@ -8,32 +8,16 @@
  ***************************************************************************/
 
 #include "stdafx.h"
-#include "librpbase/config.librpbase.h"
-#include "libromdata/config.libromdata.h"
 
 #include "Wim.hpp"
-#include "wim_structs.h"
-
-#ifdef ENABLE_XML
-// TinyXML2
-#  include "tinyxml2.h"
-using namespace tinyxml2;
-// Character set conversion, needed for TinyXML2
-// because WIM uses UTF-16LE manifests.
-#  include "librptext/conversion.hpp"
-#endif /* ENABLE_XML */
+#include "Wim_p.hpp"
 
 // Other rom-properties libraries
-#include "librpbase/timeconv.h"
 using namespace LibRpBase;
 using namespace LibRpFile;
-using namespace LibRpText;
 
 // C includes
 #include <uchar.h>
-
-// C includes (C++ namespace)
-#include <ctime>
 
 // C++ STL classes
 using std::array;
@@ -41,40 +25,6 @@ using std::string;
 using std::vector;
 
 namespace LibRomData {
-
-class WimPrivate final : public RomDataPrivate
-{
-public:
-	explicit WimPrivate(const IRpFilePtr &file);
-
-private:
-	typedef RomDataPrivate super;
-	RP_DISABLE_COPY(WimPrivate)
-
-public:
-	/** RomDataInfo **/
-	static const array<const char*, 3+1> exts;
-	static const array<const char*, 1+1> mimeTypes;
-	static const RomDataInfo romDataInfo;
-
-public:
-	// WIM header
-	WIM_Header wimHeader;
-
-	// WIM version
-	// NOTE: WIMs pre-1.13 are being detected but won't
-	// be read due to the format being different.
-	WIM_Version_Type versionType;
-
-#ifdef ENABLE_XML
-public:
-	/**
-	 * Add fields from the WIM image's XML manifest.
-	 * @return 0 on success; non-zero on error.
-	 */
-	int addFields_XML();
-#endif /* ENABLE_XML */
-};
 
 ROMDATA_IMPL(Wim)
 
@@ -95,342 +45,6 @@ const array<const char*, 1+1> WimPrivate::mimeTypes = {{
 const RomDataInfo WimPrivate::romDataInfo = {
 	"WIM", exts.data(), mimeTypes.data()
 };
-
-struct WimWindowsLanguages {
-	string language;
-	//string default_language;	//not used right now
-};
-
-struct WimWindowsVersion {
-	unsigned int majorversion = 0;
-	unsigned int minorversion = 0;
-	unsigned int buildnumber = 0;
-	unsigned int spbuildnumber = 0;
-	unsigned int splevel = 0; // only in windows 7+, added some time around build 6608-6730
-};
-
-struct WimWindowsInfo {
-	WimWindowsArchitecture arch = Wim_Arch_x86;
-	string productname;
-	string editionid;
-	//string installationtype;	// not used right now
-	//string hal;			// not used right now
-	//string producttype;		// not used right now
-	//string productsuite;		// not used right now
-	WimWindowsLanguages languages;
-	WimWindowsVersion version;
-	string systemroot;
-};
-
-struct WimIndex {
-	// if you have more than 2^32 indices in a wim
-	// you probably have bigger issues
-	uint32_t index = 0;
-	uint64_t dircount = 0;
-	uint64_t filecount = 0;
-	uint64_t totalbytes = 0;
-	//time_t creationtime = 0;	// not used right now
-	time_t lastmodificationtime = 0;
-	WimWindowsInfo windowsinfo;
-	string name, description;
-	//string flags;			// not used right now
-	string dispname, dispdescription;
-	bool containswindowsimage = false;
-};
-
-#ifdef ENABLE_XML
-
-// TinyXML2 2.2.0 (Ubuntu 16.04) doesn't have XMLElement::IntText(),
-// but it *does* have QueryIntText().
-// NOTE: IntText was originally added for 2.0.1 in commit
-// 4cd1f269c28ff2c84c454270acbb47a2c8eb0403, but was removed at some point
-// before release. Was re-added at some point between 4.0.1 and 5.0.0.
-// (Merges are making it difficult to identify the exact commit.)
-static inline int XMLElement_IntText(const XMLElement *element, int defaultValue = 0)
-{
-#ifdef HAVE_TINYXML2_XMLELEMENT_INTTEXT
-	// XMLElement::IntText(int defaultValue) is available.
-	// Use it directly.
-	return element->IntText(defaultValue);
-#else /* HAVE_TINYXML2_XMLELEMENT_INTTEXT */
-	// XMLElement::IntText(int defaultValue) is *not* available.
-	// Reimplement it here.
-	int i = defaultValue;
-	element->QueryIntText(&i);
-	return i;
-#endif /* HAVE_TINYXML2_XMLELEMENT_INTTEXT */
-}
-
-/**
- * Add fields from the WIM image's XML manifest.
- * @return 0 on success; non-zero on error.
- */
-int WimPrivate::addFields_XML() 
-{
-	if (!file || !file->isOpen()) {
-		return -EIO;
-	}
-
-	// the eighth byte of the "size" is used for flags so we have to AND it
-	static constexpr uint64_t XML_MAX_SIZE = 16U*1024U*1024U;
-	uint64_t size = (wimHeader.xml_resource.size & 0x00FFFFFFFFFFFFFFULL);
-	assert(size <= XML_MAX_SIZE);
-	if (size > XML_MAX_SIZE) {
-		// XML is larger than 16 MB, which doesn't make any sense.
-		return -ENOMEM;
-	}
-
-	// XML data is UTF-16LE, so the size should be a multiple of 2.
-	// TODO: Error out if it isn't?
-	assert(size % 2 == 0);
-	size &= ~1ULL;
-
-	// Read the WIM XML data.
-	file->rewind();
-	file->seek(wimHeader.xml_resource.offset_of_xml); 
-	// if seek is invalid
-	if (file->tell() != static_cast<off64_t>(wimHeader.xml_resource.offset_of_xml)) {
-		return -EIO;
-	}
-
-	char16_t *const xml_data = new char16_t[static_cast<size_t>(size / 2)];
-	size_t bytes_read = file->read(xml_data, static_cast<size_t>(size));
-	if (bytes_read != size) {
-		delete[] xml_data;
-		return -EIO;
-	}
-
-	// the xml inside wims are utf-16 but tinyxml only supports utf-8
-	// this means we have to do some conversion
-	string utf8_xml = LibRpText::utf16le_to_utf8(xml_data, static_cast<int>(size / 2));
-	delete[] xml_data;
-
-	XMLDocument document;
-	int xerr = document.Parse(utf8_xml.c_str(), utf8_xml.size());
-	if (xerr != XML_SUCCESS) {
-		return 3;
-	}
-
-	const XMLElement *const wim_element = document.FirstChildElement("WIM");
-	if (!wim_element) {
-		// No wim element.
-		// TODO: Better error code.
-		return -EIO;
-	}
-
-	// Sanity check: Allow up to 256 images.
-	unsigned int number_of_images = wimHeader.number_of_images;
-	assert(number_of_images <= 256);
-	if (number_of_images > 256)
-		number_of_images = 256;
-
-	vector<WimIndex> images;
-	images.reserve(number_of_images);
-	const XMLElement *currentimage = wim_element->FirstChildElement("IMAGE");
-
-	const char *const s_unknown = C_("Wim", "(unknown)");
-	for (uint32_t i = 0; i < wimHeader.number_of_images; i++) {
-		assert(currentimage != nullptr);
-		if (!currentimage)
-			break;
-
-		WimIndex currentindex;
-		currentindex.index = i + 1;
-
-		// the last modification time is split into a high part and a
-		// low part so we shift and add them together
-		const XMLElement *const creationTime = currentimage->FirstChildElement("CREATIONTIME");
-		if (creationTime) {
-			const XMLElement *const highPart = creationTime->FirstChildElement("HIGHPART");
-			const XMLElement *const lowPart = creationTime->FirstChildElement("LOWPART");
-			if (highPart && lowPart) {
-				const char *const s_highPart = highPart->GetText();
-				const char *const s_lowPart = lowPart->GetText();
-				if (s_highPart && s_lowPart) {
-					// Parse HIGHPART and LOWPART, then combine them like FILETIME.
-					const uint32_t lastmodtime_high = strtoul(s_highPart, nullptr, 16);
-					const uint32_t lastmodtime_low = strtoul(s_lowPart, nullptr, 16);
-					currentindex.lastmodificationtime = WindowsSplitTimeToUnixTime(lastmodtime_high, lastmodtime_low);
-				}
-			}
-		}
-
-		const XMLElement *const windowsinfo = currentimage->FirstChildElement("WINDOWS");
-		if (windowsinfo) {
-			currentindex.containswindowsimage = true;
-			const XMLElement *const arch = windowsinfo->FirstChildElement("ARCH");
-			if (arch) {
-				currentindex.windowsinfo.arch = static_cast<WimWindowsArchitecture>(XMLElement_IntText(arch, 0));
-			}
-
-			const XMLElement *const editionId = windowsinfo->FirstChildElement("EDITIONID");
-			if (editionId) {
-				const char *s_editionId = windowsinfo->FirstChildElement("EDITIONID")->GetText();
-				currentindex.windowsinfo.editionid = (s_editionId ? s_editionId : s_unknown);
-			} else {
-				currentindex.windowsinfo.editionid = s_unknown;
-			}
-
-			const XMLElement *const languages = windowsinfo->FirstChildElement("LANGUAGES");
-			if (languages) {
-				// NOTE: Only retrieving the first language.
-				const XMLElement *const language = languages->FirstChildElement("LANGUAGE");
-				if (language) {
-					const char *s_language = language->GetText();
-					if (s_language) {
-						currentindex.windowsinfo.languages.language = s_language;
-					}
-				}
-			}
-			if (currentindex.windowsinfo.languages.language.empty()) {
-				currentindex.windowsinfo.languages.language = s_unknown;
-			}
-
-			const XMLElement *const version = windowsinfo->FirstChildElement("VERSION");
-			if (version) {
-				const XMLElement *const ver_major = version->FirstChildElement("MAJOR");
-				if (ver_major) {
-					currentindex.windowsinfo.version.majorversion = XMLElement_IntText(ver_major, 0);
-				}
-				const XMLElement *const ver_minor = version->FirstChildElement("MINOR");
-				if (ver_minor) {
-					currentindex.windowsinfo.version.minorversion = XMLElement_IntText(ver_minor, 0);
-				}
-				const XMLElement *const ver_build = version->FirstChildElement("BUILD");
-				if (ver_build) {
-					currentindex.windowsinfo.version.buildnumber = XMLElement_IntText(ver_build, 0);
-				}
-				const XMLElement *const ver_spbuild = version->FirstChildElement("SPBUILD");
-				if (ver_spbuild) {
-					currentindex.windowsinfo.version.spbuildnumber = XMLElement_IntText(ver_spbuild, 0);
-				}
-			}
-		} else {
-			currentindex.containswindowsimage = false;
-		}
-
-		// some wims don't have these fields, so we
-		// need to set up fallbacks - the hierarchy goes
-		// display name -> name -> "(None)"
-		const char *const s_none = C_("Wim", "(none)");
-		const XMLElement *const elem_name = currentimage->FirstChildElement("NAME");
-		if (elem_name) {
-			const char *const s_name = elem_name->GetText();
-			currentindex.name = (s_name ? s_name : s_none);
-		}
-		const XMLElement *const elem_desc = currentimage->FirstChildElement("DESCRIPTION");
-		if (elem_desc) {
-			const char *const s_desc = elem_desc->GetText();
-			currentindex.description = (s_desc ? s_desc : s_none);
-		}
-		const XMLElement *const elem_dispName = currentimage->FirstChildElement("DISPLAYNAME");
-		if (elem_dispName) {
-			const char *const s_dispName = elem_dispName->GetText();
-			currentindex.dispname = (s_dispName ? s_dispName : s_none);
-		}
-		const XMLElement *const elem_dispDesc = currentimage->FirstChildElement("DISPLAYDESCRIPTION");
-		if (elem_dispDesc) {
-			const char *const s_dispDesc = elem_dispDesc->GetText();
-			currentindex.dispdescription = (s_dispDesc ? s_dispDesc : s_none);
-		}
-
-		images.push_back(std::move(currentindex));
-		currentimage = currentimage->NextSiblingElement();
-	}
-
-	auto *const vv_data = new RomFields::ListData_t();
-	vv_data->reserve(number_of_images);
-
-	// loop for the rows
-	unsigned int idx = 1;
-	for (const auto &image : images) {
-		vv_data->resize(vv_data->size()+1);
-		auto &data_row = vv_data->at(vv_data->size()-1);
-		data_row.reserve(10);
-		data_row.push_back(fmt::to_string(idx++));
-		data_row.push_back(image.name);
-		data_row.push_back(image.description);
-		data_row.push_back(image.dispname);
-		data_row.push_back(image.dispdescription);
-
-		// Pack the 64-bit time_t into a string.
-		RomFields::TimeString_t time_string;
-		time_string.time = image.lastmodificationtime;
-		data_row.emplace_back(time_string.str, sizeof(time_string.str));
-
-		if (image.containswindowsimage == false) {
-			// No Windows image. Add empty strings to complete the row.
-			for (unsigned int i = 4; i > 0; i--) {
-				data_row.emplace_back();
-			}
-			continue;
-		}
-
-		const auto &windowsver = image.windowsinfo.version;
-		data_row.push_back(
-			fmt::format(FSTR("{:d}.{:d}.{:d}.{:d}"),
-				windowsver.majorversion, windowsver.minorversion,
-				windowsver.buildnumber, windowsver.spbuildnumber));
-
-		const auto &windowsinfo = image.windowsinfo;
-		data_row.push_back(windowsinfo.editionid);
-		const char *archstring;
-		switch (windowsinfo.arch) {
-			default:
-				archstring = nullptr;
-				break;
-			case Wim_Arch_x86:
-				archstring = "x86";
-				break;
-			case Wim_Arch_ARM32:
-				archstring = "ARM32";
-				break;
-			case Wim_Arch_IA64:
-				archstring = "IA64";
-				break;
-			case Wim_Arch_AMD64:
-				archstring = "x64";
-				break;
-			case Wim_Arch_ARM64:
-				archstring = "ARM64";
-				break;
-		}
-		if (archstring) {
-			data_row.emplace_back(archstring);
-		} else {
-			data_row.push_back(fmt::format(FRUN(C_("RomData", "Unknown ({:d})")),
-				static_cast<int>(windowsinfo.arch)));
-		}
-		data_row.push_back(windowsinfo.languages.language);
-	}	
-
-	static const array<const char*, 10> field_names = {{
-		NOP_C_("Wim|Images", "#"),
-		NOP_C_("Wim|Images", "Name"),
-		NOP_C_("Wim|Images", "Description"),
-		NOP_C_("Wim|Images", "Display Name"),
-		NOP_C_("Wim|Images", "Display Desc."),
-		NOP_C_("Wim|Images", "Last Modified"),
-		NOP_C_("Wim|Images", "OS Version"),
-		NOP_C_("Wim|Images", "Edition"),
-		NOP_C_("Wim|Images", "Architecture"),
-		NOP_C_("Wim|Images", "Language"),
-	}};
-	vector<string> *const v_field_names = RomFields::strArrayToVector_i18n("Wim|Images", field_names);
-
-	RomFields::AFLD_PARAMS params;
-	params.flags = RomFields::RFT_LISTDATA_SEPARATE_ROW;
-	params.col_attrs.is_timestamp = (1U << 5);
-	params.col_attrs.dtflags = static_cast<RomFields::DateTimeFlags>(
-		RomFields::RFT_DATETIME_HAS_DATE | RomFields::RFT_DATETIME_HAS_TIME);
-	params.headers = v_field_names;
-	params.data.single = vv_data;
-	// TODO: Header alignment?
-	fields.addField_listData(C_("Wim", "Images"), &params);
-
-	return 0;
-}
-#endif /* ENABLE_XML */
 
 WimPrivate::WimPrivate(const IRpFilePtr &file)
 	: super(file, &romDataInfo)
@@ -663,7 +277,16 @@ int Wim::loadFieldData(void)
 
 #ifdef ENABLE_XML
 	// Add fields from the WIM image's XML manifest.
-	d->addFields_XML();
+	int ret = d->addFields_XML();
+	if (ret != 0) {
+		d->fields.addField_string(C_("RomData", "Warning"),
+			C_("RomData", "XML parsing failed."),
+			RomFields::STRF_WARNING);
+	}
+#else /* !ENABLE_XML */
+	d->fields.addField_string(C_("RomData", "Warning"),
+		C_("RomData", "XML parsing is disabled in this build."),
+		RomFields::STRF_WARNING);
 #endif /* ENABLE_XML */
 
 	// Finished reading the field data.